legacy-memcached-django/requirements.txt
appengine-python-standard==1.0.0
# asgiref==3.5.2
# attrs==21.4.0
# cachetools==5.2.0
# certifi==2022.6.15
# charset-normalizer==2.1.0
Django==4.0.6
# frozendict==2.3.4
# google-auth==2.9.1
# idna==3.3
# mock==4.0.3
# Pillow==9.2.0
# protobuf==4.21.3
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pytz==2022.1
# requests==2.28.1
# rsa==4.9
# ruamel.yaml==0.17.21
# ruamel.yaml.clib==0.2.6
# six==1.16.0
# sqlparse==0.4.2
# urllib3==1.26.11
google-cloud-logging
